Refactor the notify service into the Notify scope
