+++ spec/lib/tasks/notify_spec.rb
@@ -7,28 +7,28 @@ RSpec.describe "Notify task", type: :rake do
 
   describe "notify:letters:ad_renewals" do
     it "runs without error" do
-      expect(NotifyBulkAdRenewalLettersService).to receive(:run).and_return([build(:registration)])
+      expect(Notify::BulkAdRenewalLettersService).to receive(:run).and_return([build(:registration)])
       expect { subject.invoke }.not_to raise_error
     end
   end
 
   describe "notify:letters:digital_renewals" do
     it "runs without error" do
-      expect(NotifyBulkDigitalRenewalLettersService).to receive(:run).and_return([build(:registration)])
+      expect(Notify::BulkDigitalRenewalLettersService).to receive(:run).and_return([build(:registration)])
       expect { subject.invoke }.not_to raise_error
     end
   end
 
   describe "notify:test:ad_renewal_letter" do
     it "runs without error" do
-      expect(NotifyAdRenewalLetterService).to receive(:run)
+      expect(Notify::AdRenewalLetterService).to receive(:run)
       expect { subject.invoke }.not_to raise_error
     end
   end
 
   describe "notify:test:digital_renewal_letter" do
     it "runs without error" do
-      expect(NotifyDigitalRenewalLetterService).to receive(:run)
+      expect(Notify::DigitalRenewalLetterService).to receive(:run)
       expect { subject.invoke }.not_to raise_error
     end
   end
